FSDUAsigmaproject/requirements.txt
africastalking==1.2.4
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
contextlib2==21.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-session-timeout==0.1.0
et-xmlfile==1.1.0
idna==3.3
MarkupPy==1.14
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
schema==0.7.5
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
